Things work.
Bug testing still needs to be done, but I believe all the parts that
need to work in order to process robot data are functional. The next
step is to create a program that can skim data from one CompleteData
file and append it to another so that changes made are not lost after
every new data dump.